Moved common elements to utils.py file, created first pass of test executor and building test payload
